Update site fonts, header, and color scheme
The commit includes changes to the website design. It updates the font used on the site to 'Ubuntu' and 'Railway' and modifies the color scheme. The website header has also been enhanced, including the addition of site titles as clickable links. The site's title in the Hugo configuration has also been updated to 'Hugo Theme Tech Minimal'.