Set limit for posts shown on homepage
A parameter `postsPerHomePage` has been added to the Hugo configuration file, hugo.toml, which allows the control of the number of posts displayed on the home page of the site. The home.html file was updated to use this new parameter, thereby ensuring only the specified number of latest posts are shown.