Refactor social icons display and add RSS icon
The header's section for social media icons has been refactored to improve code reusability and maintainability. It now iterates over a range of social platforms defined in the site parameters. Additionally, an SVG for an RSS icon has been added. The list of social platforms in the site parameters has been expanded to include Github and RSS.