Add draft label to site posts
A new feature has been incorporated into the site that adds a 'draft' label to all drafts on site. This labelling appears on different sections of the site such as tags, single posts, home and list pages. This enhancement helps distinguish drafts from other posts and adds useful signifiers to content in development.