Refactor article presentation and update CSS rules
Moved article presentation functionality to a new "article.html" partial, which is now used in both the home and list layouts. This change simplifies the HTML code in these files and provides a centralized location for article display changes. The CSS for article elements was also updated, shifting from h2 to article tags and adjusting margins for a more balanced visual spacing.