Add Google Analytics integration
The template code for Google Analytics has been inserted into the base HTML layout. This is conditional and only loads if the site is not in server mode. To make this function, the Google Analytics ID has been added to the site configuration file (hugo.toml) under the new services section.